moved routes into folers - root and dashboard, minor changes
