--- client/src/routes/SignInPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function SignInPage() {
-  return (
-    <div>SignInPage</div>
-  )
-}
