--- client/src/routes/SignUpPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function SignUpPage() {
-  return (
-    <div>SignUpPage</div>
-  )
-}
