+++ client/src/routes/dashboard/RoomPage.jsx
@@ -2,6 +2,6 @@ import React from 'react'
 
 export default function RoomPage() {
   return (
-    <div>Room</div>
+    <div>Room PAGE</div>
   )
 }
