+++ client/src/routes/root/SignInPage.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { SignIn } from "@clerk/clerk-react";
+export default function SignInPage() {
+  return (
+    <SignIn path="/sign-in" routing="path" signUpUrl="/sign-up" />
+  )
+}
