+++ client/src/routes/root/SignUpPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import { SignUp } from "@clerk/clerk-react";
+export default function SignUpPage() {
+  return (
+    <div>SignUpPage
+        <SignUp/>
+    </div>
+  )
+}
