+++ client/src/routes/root/UserProfilePage.jsx
@@ -1,0 +1,6 @@
+import { UserProfile } from "@clerk/clerk-react";
+export default function UserProfilePage() {
+  return (
+    <UserProfile path="/dashboard/user-profile" routing="path" />
+  )
+}
